Respect indentation in the line for sub tree in BasicFormat
